# SwaggerFile.json
{
...
  "info": {
    "description": "API for the Quiz Master application",
    "termsOfService": "",
    "title": "Quiz Master API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/login": {
      "get": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "description": "Authenticate and login as an admin",
        "parameters": [
          {
            "description": "Admin username",
            "in": "formData",
            "name": "username",
            "required": true,
            "type": "string"
          },
          {
            "description": "Admin password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "admin": {
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "username": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Login successful"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Login failed"
          }
        },
        "summary": "Admin Login",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "description": "Authenticate and login as an admin",
        "parameters": [
          {
            "description": "Admin username",
            "in": "formData",
            "name": "username",
            "required": true,
            "type": "string"
          },
          {
            "description": "Admin password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "admin": {
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "username": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Login successful"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Login failed"
          }
        },
        "summary": "Admin Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/login": {
      "get": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "description": "Authenticate and login a user",
        "parameters": [
          {
            "description": "User email",
            "in": "formData",
            "name": "email",
            "required": true,
            "type": "string"
          },
          {
            "description": "User password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Login successful"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Login failed"
          }
        },
        "summary": "User Login",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "description": "Authenticate and login a user",
        "parameters": [
          {
            "description": "User email",
            "in": "formData",
            "name": "email",
            "required": true,
            "type": "string"
          },
          {
            "description": "User password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Login successful"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Login failed"
          }
        },
        "summary": "User Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/register": {
      "get": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "description": "Register a new user account",
        "parameters": [
          {
            "description": "User full name",
            "in": "formData",
            "name": "name",
            "required": true,
            "type": "string"
          },
          {
            "description": "User email",
            "in": "formData",
            "name": "email",
            "required": true,
            "type": "string"
          },
          {
            "description": "User password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          },
          {
            "description": "Password confirmation",
            "in": "formData",
            "name": "confirm_password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Registration successful"
          },
          "400": {
            "description": "Invalid input or user already exists"
          }
        },
        "summary": "User Registration",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "application/json"
        ],
        "description": "Register a new user account",
        "parameters": [
          {
            "description": "User full name",
            "in": "formData",
            "name": "name",
            "required": true,
            "type": "string"
          },
          {
            "description": "User email",
            "in": "formData",
            "name": "email",
            "required": true,
            "type": "string"
          },
          {
            "description": "User password",
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          },
          {
            "description": "Password confirmation",
            "in": "formData",
            "name": "confirm_password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Registration successful"
          },
          "400": {
            "description": "Invalid input or user already exists"
          }
        },
        "summary": "User Registration",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "sessionAuth": {
      "in": "cookie",
      "name": "session",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}